chore(Func_Jobs): modernize funciton to use latest from go fn SDK
