+++ functions/Func_Jobs/api/audit.go
@@ -2,88 +2,45 @@ package api
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"strconv"
 	"strings"
 	"sync"
 
-	fdk "github.com/CrowdStrike/foundry-fn-go"
 	"github.com/Crowdstrike/foundry-sample-rapid-response/functions/Func_Jobs/api/models"
 	"github.com/crowdstrike/gofalcon/falcon/client"
+
+	fdk "github.com/CrowdStrike/foundry-fn-go"
 )
 
-// AuditsHandler executes a given request to the FaaS function.
-type AuditsHandler struct {
-	conf *models.Config
-}
-
-// NewAuditsHandler returns an initialized instance of the AuditsHandler.
-func NewAuditsHandler(conf *models.Config) *AuditsHandler {
-	return &AuditsHandler{
-		conf: conf,
-	}
-}
-
-func (h *AuditsHandler) Handle(ctx context.Context, request fdk.Request) fdk.Response {
-	response := fdk.Response{}
-	var err error
-
-	fc, err := models.FalconClient(ctx, h.conf, request)
-	if err != nil {
-		response.Code = http.StatusInternalServerError
-		response.Errors = append(response.Errors, fdk.APIError{Code: http.StatusBadRequest, Message: "fail to initialize client"})
-		return response
-	}
-
-	audits, errs := h.auditDetails(ctx, &request, fc)
+func auditHandlerFn(ctx context.Context, r fdk.Request) fdk.Response {
+	audits, errs := auditDetails(ctx, &r, getFalconClient(ctx))
 	if len(errs) != 0 {
-		response.Code = http.StatusInternalServerError
-		response.Errors = errs
-		return response
+		return fdk.ErrResp(errs...)
 	}
-
-	body, err := json.Marshal(audits)
-	if err != nil {
-		response.Code = http.StatusInternalServerError
-		response.Errors = []fdk.APIError{models.NewAPIError(http.StatusInternalServerError, fmt.Sprintf("failed to marshal the response body with err:%v", err))}
-		return response
-	}
-	response.Body = json.RawMessage(body)
-	response.Code = http.StatusOK
-	return response
+	return fdk.Response{Code: http.StatusOK, Body: fdk.JSON(audits)}
 }
 
-// jobInfo gets all the jobs for the app.
-func (h *AuditsHandler) auditDetails(ctx context.Context, request *fdk.Request, fc *client.CrowdStrikeAPISpecification) (*models.AuditResponse, []fdk.APIError) {
-	// Default limit set to 10
-	limit := 10
-	var response models.AuditResponse
-	var errs []fdk.APIError
-	var err error
-
-	response.Resources = make([]models.Audit, 0)
-	response.Meta = &models.Paging{
-		Limit: limit,
+func auditDetails(ctx context.Context, request *fdk.Request, fc *client.CrowdStrikeAPISpecification) (*models.AuditResponse, []fdk.APIError) {
+	response := models.AuditResponse{
+		Resources: make([]models.Audit, 0),
+		Meta:      &models.Paging{Limit: 10},
 	}
 
 	limitParam := request.Params.Query.Get(queryLimit)
 	if limitParam != "" {
-		limit, err = strconv.Atoi(limitParam)
+		limit, err := strconv.Atoi(limitParam)
 		if err != nil {
-			return &response, []fdk.APIError{{
+			return nil, []fdk.APIError{{
 				Code:    http.StatusBadRequest,
 				Message: fmt.Sprintf("limit is not an integer: %v", err),
 			}}
-
 		}
+		response.Meta.Limit = limit
 	}
 
-	nOffset := request.Params.Query.Get(queryPrevOffset)
-	qOffset := request.Params.Query.Get(queryNextOffset)
 	filter := request.Params.Query.Get(queryParamFilter)
-
 	if filter != "" {
 		pair := strings.Split(filter, ":")
 		key := pair[0]
@@ -96,6 +53,8 @@ func (h *AuditsHandler) auditDetails(ctx context.Context, request *fdk.Request, 
 		filter = pair[1]
 	}
 
+	nOffset := request.Params.Query.Get(queryPrevOffset)
+	qOffset := request.Params.Query.Get(queryNextOffset)
 	if qOffset != "" && nOffset != "" {
 		return &response, []fdk.APIError{{
 			Code:    http.StatusBadRequest,
@@ -140,8 +99,11 @@ func (h *AuditsHandler) auditDetails(ctx context.Context, request *fdk.Request, 
 			Message: fmt.Errorf("error constructing FQL sort: %s", err.Error()).Error(),
 		}}
 	}
+
+	limit := response.Meta.Limit
+
 	searchReq := models.SearchObjectsRequest{
-		Collection: h.conf.AuditLogsCollection,
+		Collection: collectionAudit,
 		Limit:      limit,
 		Offset:     offset,
 		Sort:       fqlSort,
@@ -162,7 +124,7 @@ func (h *AuditsHandler) auditDetails(ctx context.Context, request *fdk.Request, 
 		Count: totalRemain,
 	}
 	if totalRemain == 0 {
-		return &response, errs
+		return &response, nil
 	}
 
 	errChan := make(chan error, limit)
@@ -175,20 +137,21 @@ func (h *AuditsHandler) auditDetails(ctx context.Context, request *fdk.Request, 
 		go func(id string, count int) {
 			defer wg.Done()
 
-			audit, errs := auditInfo(ctx, id, h.conf, fc)
+			audit, errs := getObject[models.Audit](ctx, id, collectionAudit, fc)
 			if len(errs) != 0 {
 				var jobGetErr error
 				jobGetErr = fmt.Errorf("failed to get job: %s id: err: %v", id, errs)
 				errChan <- jobGetErr
 				return
 			}
-			auditDetail[count] = audit
+			auditDetail[count] = &audit
 		}(id, idx)
 	}
 
 	wg.Wait()
 	close(errChan)
 
+	var errs []fdk.APIError
 	for err := range errChan {
 		errs = append(errs, models.NewAPIError(http.StatusInternalServerError, err.Error()))
 	}
